Fix usage of timeval_ms()
First, fix the timeval_ms() implementation to not have K&R but ANSI
argument semantics by adding a missing void.

timeval_ms() returns an int64_t, not uint64_t or long long. Consistently
use int64_t for variables and PRI*64 as format string.

While at it, change a few related variables to bool for clarity.

Note that timeval_ms() may return a negative error code, but not a
single caller checks for that.

Reviewed-on: http://openocd.zylin.com/3499
Tested-by: jenkins

+++ src/flash/nor/aduc702x.c
@@ -371,7 +371,7 @@ static int aduc702x_check_flash_completion(struct target *target, unsigned int t
 {
 	uint8_t v = 4;
 
-	long long endtime = timeval_ms() + timeout_ms;
+	int64_t endtime = timeval_ms() + timeout_ms;
 	while (1) {
 		target_read_u8(target, ADUC702x_FLASH + ADUC702x_FLASH_FEESTA, &v);
 		if ((v & 4) == 0)
