+++ src/flash/nor/atsamv.c
@@ -147,7 +147,7 @@ static int samv_efc_perform_command(struct target *target,
 {
 	int r;
 	uint32_t v;
-	long long ms_now, ms_end;
+	int64_t ms_now, ms_end;
 
 	if (status)
 		*status = 0;
