+++ src/helper/log.c
@@ -45,10 +45,10 @@ int debug_level = -1;
 static FILE *log_output;
 static struct log_callback *log_callbacks;
 
-static long long last_time;
-static long long current_time;
+static int64_t last_time;
+static int64_t current_time;
 
-static long long start;
+static int64_t start;
 
 static const char * const log_strings[5] = {
 	"User : ",
@@ -134,12 +134,12 @@ static void log_puts(enum log_levels level,
 	if (strlen(string) > 0) {
 		if (debug_level >= LOG_LVL_DEBUG) {
 			/* print with count and time information */
-			int t = (int)(timeval_ms()-start);
+			int64_t t = timeval_ms() - start;
 #ifdef _DEBUG_FREE_SPACE_
 			struct mallinfo info;
 			info = mallinfo();
 #endif
-			fprintf(log_output, "%s%d %d %s:%d %s()"
+			fprintf(log_output, "%s%d %" PRId64 " %s:%d %s()"
 #ifdef _DEBUG_FREE_SPACE_
 				" %d"
 #endif
@@ -410,12 +410,12 @@ void keep_alive()
 		if (gdb_actual_connections)
 			LOG_WARNING("keep_alive() was not invoked in the "
 				"1000ms timelimit. GDB alive packet not "
-				"sent! (%lld). Workaround: increase "
+				"sent! (%" PRId64 "). Workaround: increase "
 				"\"set remotetimeout\" in GDB",
 				current_time-last_time);
 		else
 			LOG_DEBUG("keep_alive() was not invoked in the "
-				"1000ms timelimit (%lld). This may cause "
+				"1000ms timelimit (%" PRId64 "). This may cause "
 				"trouble with GDB connections.",
 				current_time-last_time);
 	}
